// packages/client/src/constants/BetResult.ts
export enum BetResult {
  Win = 'win',
  Returned = 'returned',
  Lost = 'lost',
  CashOut = 'cashout',  
  UnSettled = 'unsettled',
  UnSettledPrice = 'unsettledprice',
  All = 'All',
}

export enum BetType {
  All = 'All',
  Single = 'Single',
  Multiple = 'Multiple',
  System = 'System',
  Chain = 'Chain',
}